Backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Илья Смоленский",email = "smolenskiy20030409@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.11.14)",
    "aiosignal (==1.3.2)",
    "aiosmtplib (==3.0.2)",
    "aiosqlite (==0.21.0)",
    "alembic (==1.16.1)",
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "async-timeout (==5.0.1)",
    "asyncpg (==0.30.0)",
    "attrs (==25.3.0)",
    "bcrypt (==4.3.0)",
    "blinker (==1.9.0)",
    "certifi (==2025.1.31)",
    "cffi (==1.17.1)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "cryptography (==45.0.4)",
    "dnspython (==2.7.0)",
    "ecdsa (==0.19.1)",
    "email-validator (==2.2.0)",
    "exceptiongroup (==1.2.2)",
    "fastapi (==0.115.12)",
    "fastapi-cli (==0.0.7)",
    "fastapi-mail (==1.5.0)",
    "frozenlist (==1.5.0)",
    "greenlet (==3.2.2)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "iniconfig (==2.1.0)",
    "jinja2 (==3.1.6)",
    "mako (==1.3.10)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "multidict (==6.2.0)",
    "packaging (==25.0)",
    "passlib (==1.7.4)",
    "pluggy (==1.6.0)",
    "propcache (==0.3.1)",
    "psycopg2-binary (==2.9.10)",
    "pyasn1 (==0.6.1)",
    "pycparser (==2.22)",
    "pydantic (==2.10.6)",
    "pydantic-settings (==2.9.1)",
    "pydantic-core (==2.27.2)",
    "pygments (==2.19.1)",
    "pytest (==8.4.0)",
    "python-dotenv (==1.1.0)",
    "python-jose (==3.5.0)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.2)",
    "rich (==13.9.4)",
    "rich-toolkit (==0.13.2)",
    "rsa (==4.9.1)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.41)",
    "starlette (==0.46.1)",
    "tomli (==2.2.1)",
    "typer (==0.15.2)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.13.0)",
    "uvicorn (==0.34.0)",
    "watchfiles (==1.0.4)",
    "websockets (==15.0.1)",
    "yarl (==1.18.3)",
    "pytest-asyncio (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "app" }]

[tool.poetry.scripts]
start = "app.app:run"
